Add assertion failure messages.
When there are more than 1 assertion in a test method, a failure
message helps to identify which assertion failed.

Once an assertion fails, any other assertions after it in the
same test method will not run.

Thus, the message helps to know which one of the assertions failed.
This can help with debugging process.